pyproject.toml
[project]
name = "hdr_analysis"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "SiTinc", email = "23432137+Tinnci@users.noreply.github.com"},
]
dependencies = [
    "opencv-python>=4.10.0.84",
    "numpy>=2.1.2",
    "matplotlib>=3.9.2",
    "scikit-image>=0.24.0",
    "scipy>=1.14.1",
    "pandas>=2.2.3",
    "Pillow>=11.0.0",
    "tqdm>=4.66.5",
    "psutil>=6.1.0",
    "pyqt6>=6.7.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
